fix: endpoint to get created exams
