+++ apps/api/src/controllers/exams.ts
@@ -225,26 +225,3 @@ export const compareParticipantAnswers = async (
     });
   } catch (error) {}
 };
-
-export const getExamsCreatedBySpecificUser = async (
-  req: Request,
-  res: Response
-) => {
-  const userAddress = req.params.address;
-
-  try {
-    const exams = await prisma.exam.findMany({
-      where: {
-        creatorAddress: {
-          mode: 'insensitive',
-          equals: userAddress,
-        },
-      },
-    });
-
-    res.statusCode = 200;
-    res.send({
-      exams,
-    });
-  } catch (e) {}
-};
