+++ apps/api/src/routes/users.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+import { getExamsCreatedBySpecificUser } from '../controllers/users';
+
+export const router = express.Router();
+router.route('/:address/exams').get(getExamsCreatedBySpecificUser);
